requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.0
future==1.0.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
kazoo==2.5.0
MarkupSafe==3.0.2
packaging==24.2
paho-mqtt==2.1.0
pluggy==1.5.0
pykafka==2.8.0
pytest==8.3.4
python-dotenv==1.0.1
python-engineio==4.11.1
python-socketio==5.12.0
PyYAML==6.0.2
requests==2.32.3
serial==0.0.97
setuptools==75.6.0
simple-websocket==1.1.0
six==1.17.0
tabulate==0.9.0
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
